+++ .ci/monolithic-windows.sh
@@ -37,7 +37,7 @@ function at-exit {
   shopt -s nullglob
   if command -v buildkite-agent 2>&1 >/dev/null
   then
-    python "${MONOREPO_ROOT}"/.ci/generate_test_report.py ":windows: Windows x64 Test Results" \
+    python "${MONOREPO_ROOT}"/.ci/generate_test_report_buildkite.py ":windows: Windows x64 Test Results" \
       "windows-x64-test-results" $retcode "${BUILD_DIR}"/test-results.*.xml
   fi
 }
